requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
python-decouple==3.8
sqlparse==0.5.3
pytz==2024.1
tzdata==2024.1
asgiref==3.8.1
PyMySQL==1.1.1
gunicorn==21.2.0
drf-spectacular==0.28.0
cryptography>=42.0.0
channels==4.2.2

# IA - Google Gemini
google-auth==2.39.0
google-genai==1.11.0
cachetools==5.5.2
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
urllib3==2.2.2
requests==2.32.3
certifi==2024.7.4
idna==3.7
charset-normalizer==3.3.2

# Extras útiles (opcional)
pandas==2.2.2
openpyxl==3.1.5
numpy==2.0.1
python-dateutil==2.9.0.post0
packaging==24.1

# Core para WebSockets
daphne==4.2.0
autobahn==24.4.2
twisted==25.5.0
txaio==23.1.1
zope-interface==7.2
hyperlink==21.0.0
incremental==24.7.2
automat==25.4.16
constantly==23.10.4
service-identity==24.2.0
pyopenssl==25.1.0

#testing
pytest==8.4.1
pytest-django==4.11.1
pytest-factoryboy==2.7.0
coverage==7.9.1